[MLIR] Introduce the type inference pass.
Type Inference uses the existing forward dataflow analysis mechanisms, with a
lattice for the types, and the `join` relationship.

It is designed to be safe to run on existing IR: it will not modify types
wherever uninstructed.
Interfaces `AllowsInputTypesRefinementInterface` and
`AllowsOuptputTypesRefinementInterface` allow fine-grained control over what
input/output types can be specialized. The type inference pass will "adapt"
types to satisfy the specified constraints. By default, this means inserting
`relax_type` and `specialize_type` ops. But that is configurable via the pass
constructor.

Notes:
* Due to potentially arbitrary constraints on block argument types coming from
  parent blocks' terminators, block argument types are never specialized.
  Instead, they are "adapted" when appropriate (using `specialize_type` ops by
  default).
* Support for input types refinement is fully controlled by
  `AllowsInputTypesRefinementInterface`.
* Support for ouptut types refinement is explicit if
  `AllowsOuptputTypesRefinementInterface` is implemented. If not, it is
  implicit if `InferTypeOpInterface` is implemented.

Differential Revision: https://reviews.llvm.org/D110768

test configurable